Flask form deleted dataclass used

+++ app/cardf.py
@@ -1,9 +1,9 @@
-from flask import (Blueprint, flash, render_template, request, url_for, abort, redirect)
+from flask import (Blueprint, flash, render_template, request, url_for, abort, redirect, request)
 from flask_login import login_required, current_user
 
 from datetime import datetime
 from .models import db
-from app.forms.card_form import CardForm
+from app.schema.card import CardSchema
 import random
 from app.models.user import User
 from app.models.card import Cards
@@ -14,7 +14,6 @@ card_bp = Blueprint('cards', __name__)
 @card_bp.route('/')
 @login_required
 def index():
-
     # gets user info
     user = User.query.filter_by(id=current_user.id).first()
     cards = user.cards.all()
@@ -33,26 +32,26 @@ def index():
         card_id = 0
         dif_tops = 0
         total_cards = 0
-    return render_template('/index.html', topic=topic,dif_tops=dif_tops, question=question, card_id=card_id, \
+    return render_template('/index.html', topic=topic, dif_tops=dif_tops, question=question, card_id=card_id, \
                            total_cards=total_cards)
 
 
 @card_bp.route('/create_card', methods=('GET', 'POST'))
 @login_required
 def create_card():
-    user = User.query.filter_by(id=current_user.id).first()
-    form = CardForm()
-    if request.method == 'POST':
-        if form.validate_on_submit():
 
-            new_card = Cards(topic=form.topic.data, question=form.question.data, author=user, typ=form.typ.data)
+    if request.method == 'POST':
+        card_schema = CardSchema(topic=request.form['topic'], question=request.form['question'], typ=request.form['typ'])
+        if card_schema.validate_form():
+            new_card = Cards(topic=card_schema.topic, question=card_schema.topic, author=current_user, \
+                             typ=card_schema.typ)
             db.session.add(new_card)
             db.session.commit()
             return redirect(url_for('cards.index'))
         else:
-            flash('invalid form')
+            return 'Invalid form', 400
 
-    return render_template('/create_card.html', form=form)
+    return render_template('/create_card.html')
 
 
 def get_post(id):
@@ -85,28 +84,31 @@ def show_cards(card_type=None):
 @card_bp.route('/<int:card_id>/update', methods=('GET', 'POST'))
 @login_required
 def update_card(card_id=None):
-    card = Cards.query.filter_by(id=card_id).first()
+    card = get_post(card_id)
     if card is None:
-        flash('Card not found')
-    form = CardForm()
+        return 'Card Not found', 400
     if request.method == 'POST':
-        if form.validate_on_submit():
-            card.topic = form.topic.data
-            card.question = form.question.data
-            card.typ = form.typ.data
+        card_schema = CardSchema(request.form['topic'], request.form['question'], request.form['typ'])
+        if card_schema.validate_form():
+            card.topic = card_schema.topic
+            card.question = card_schema.question
+            card.typ = card_schema.typ
             card.updated_time = datetime.utcnow()
             db.session.commit()
             return redirect(url_for('cards.index'))
         else:
-            flash('invalid form')
+            return "Invalid form", 400
 
-    return render_template('/create_card.html', form=form, topic=card.topic, question=card.question, typ=card.typ)
+    return render_template('/create_card.html', topic=card.topic, question=card.question, typ=card.typ)
 
 
-@card_bp.route('/<int:card_id>/delete', methods=('GET', 'POST'))
+# TO-DO: use POST method
+@card_bp.route('/<int:card_id>/delete', methods=['GET'])
 @login_required
 def delete_card(card_id=None):
     card = get_post(card_id)
+    if card is None:
+        return 'Card Not found', 400
     db.session.delete(card)
     db.session.commit()
-    return redirect(url_for('cards.index'))
+    return redirect(url_for('cards.show_cards'))
